feat: changed logger zap
