+++ server/main.go
@@ -1,13 +1,9 @@
 package main
 
 import (
-	"chatting/config"
-	"chatting/logger"
 	pubsub2 "chatting/pubsub"
-	"context"
 	"github.com/google/uuid"
 	"github.com/labstack/echo/v4"
-	log2 "github.com/labstack/gommon/log"
 	"log"
 	"net/http"
 )
@@ -25,42 +21,13 @@ func serveHome(w http.ResponseWriter, r *http.Request) {
 	http.ServeFile(w, r, "home.html")
 }
 
-const (
-	logLevelDebug = "debug"
-	logLevelInfo  = "info"
-	logLevelWarn  = "warn"
-	logLevelError = "error"
-)
-
-var Ctx context.Context
 var serverId uuid.UUID
 var pubsub pubsub2.PubSub
 
 var e *echo.Echo
 
-func init() {
-	e = echo.New()
-
-	logConfig := config.Config().GetStringMapString("logger")
-	switch logConfig["level"] {
-	case logLevelDebug:
-		e.Logger.SetLevel(log2.DEBUG)
-		break
-	case logLevelInfo:
-		e.Logger.SetLevel(log2.INFO)
-		break
-	case logLevelWarn:
-		e.Logger.SetLevel(log2.WARN)
-		break
-	case logLevelError:
-		e.Logger.SetLevel(log2.ERROR)
-		break
-	}
-
-	logger.SetLogger(e.Logger)
-}
-
 func main() {
+	e = echo.New()
 	hub := NewHub()
 	serverId = uuid.New()
 	e.Logger.Infof("server id : [%s]", serverId.String())
